implement for reminder
