+++ src/index.ts
@@ -1,51 +1,27 @@
-import { getGameInfo, organizeGameInfo, postToSlack, organizeGameInfoForLine, getTomorrowGarbageInfo, getGarbageInfoList } from "./lib";
+import {
+  LineRequestPayload,
+  LineRequestPostDataContent
+} from "./types";
 
-const LINE_CHANNEL_ACCESS_TOKEN_NPB = PropertiesService.getScriptProperties().getProperty("LINE_CHANNEL_ACCESS_TOKEN_NPB");
 const LINE_CHANNEL_ACCESS_TOKEN_HOME = PropertiesService.getScriptProperties().getProperty("LINE_CHANNEL_ACCESS_TOKEN_HOME");
 const LINE_USER_ID = PropertiesService.getScriptProperties().getProperty("LINE_USER_ID");
-const LINE_GROUP_ID = PropertiesService.getScriptProperties().getProperty("LINE_GROUP_ID");
 const LINE_API_PUSH_MESSAGE_URL = "https://api.line.me/v2/bot/message/push";
-const LINE_API_ADD_RICHMENU_URL = "https://api.line.me/v2/bot/richmenu";
 
-/**
- * Slack へ NPB の試合一覧を通知する関数
- */
-function notifyToSlackAboutNPB() {
+const validUserIdList = [LINE_USER_ID];
+// const validUserIdList = [];
+
+function sendErrorResponse(userId: string) {
+  const url = LINE_API_PUSH_MESSAGE_URL;
+  const body = "Invalid Request";
   try {
-    const gameInfo = getGameInfo();
-    console.log(gameInfo);
-    const organizedGameInfo = organizeGameInfo(gameInfo);
-    console.log(organizedGameInfo);
-    postToSlack(organizedGameInfo);
-  } catch (error) {
-    console.log(error);
-  }
-}
-
-/**
- * LINE への通知
- */
-function notifyToLINEAboutNPB() {
-  try {
-    const gameInfo = getGameInfo();
-    console.log(gameInfo);
-    const organizedGameInfoForLine = organizeGameInfoForLine(gameInfo);
-    console.log(organizedGameInfoForLine);
-    // LINE Messaging APIの利用のための下準備
-    const url = LINE_API_PUSH_MESSAGE_URL;
-    // メッセージ本文を格納する変数
-    const body = organizedGameInfoForLine.length > 0
-      ? organizedGameInfoForLine
-      : '今日は試合がありません';
-
     UrlFetchApp.fetch(url, {
       'headers': {
           'Content-Type': 'application/json; charset=UTF-8',
-          'Authorization': 'Bearer ' + LINE_CHANNEL_ACCESS_TOKEN_NPB,
+          'Authorization': 'Bearer ' + LINE_CHANNEL_ACCESS_TOKEN_HOME,
       },
       'method': 'POST',
       'payload': JSON.stringify({
-          'to': LINE_USER_ID,
+          'to': userId,
           'messages': [{
               'type': 'text',
               'text': body,
@@ -57,16 +33,10 @@ function notifyToLINEAboutNPB() {
   }
 }
 
-/**
- * ゴミ出しの通知
- */
-function notifyToLINEAboutGarbage() {
+function sendOkResponse(userId: string) {
+  const url = LINE_API_PUSH_MESSAGE_URL;
+  const body = "Valid Request";
   try {
-    const url = LINE_API_PUSH_MESSAGE_URL;
-    const tomorrowInfo = getTomorrowGarbageInfo();
-    const garbageInfoList = getGarbageInfoList();
-    const message = tomorrowInfo + garbageInfoList;
-
     UrlFetchApp.fetch(url, {
       'headers': {
           'Content-Type': 'application/json; charset=UTF-8',
@@ -74,15 +44,28 @@ function notifyToLINEAboutGarbage() {
       },
       'method': 'POST',
       'payload': JSON.stringify({
-          'to': LINE_GROUP_ID,
+          'to': userId,
           'messages': [{
               'type': 'text',
-              'text': message,
+              'text': body,
           }]
       })
-    });
+    })
   } catch (error) {
     console.log(error);
   }
 }
 
+function doPost(e: LineRequestPayload){
+  const contents: LineRequestPostDataContent = JSON.parse(e.postData.contents);
+  const userId = contents.events[0].source.userId;
+  const groupId = contents.events[0].source.groupId;
+
+  if (!validUserIdList.includes(userId)) {
+    console.log("invalid user");
+    // sendErrorResponse(userId);
+  } else {
+    console.log("valid user");
+    sendOkResponse(userId);
+  }
+}
